requirements.txt
# Core
google-adk>=0.1.0
python-dotenv>=1.0.0

# Google Cloud
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.13.0
pyarrow>=12.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0

# Visualization
plotly>=5.15.0

# Web & API
requests>=2.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
aiohttp>=3.8.0

# Dashboard
streamlit>=1.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
pylint>=3.0.0

# Development
pre-commit>=3.3.0
jupyter>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0